+++ OndselSolver/EulerArray.h
@@ -17,8 +17,8 @@ namespace MbD {
     {
         //
     public:
-        EulerArray(int count) : FullColumn<T>(count) {}
-        EulerArray(int count, const T& value) : FullColumn<T>(count, value) {}
+        EulerArray(size_t count) : FullColumn<T>(count) {}
+        EulerArray(size_t count, const T& value) : FullColumn<T>(count, value) {}
         EulerArray(std::initializer_list<T> list) : FullColumn<T>{ list } {}
         void initialize() override;
         void equalFullColumn(FColsptr<T> fullCol);
