+++ OndselSolver/FullRow.h
@@ -33,8 +33,8 @@ namespace MbD {
 	public:
 		FullRow() : FullVector<T>() {}
 		FullRow(std::vector<T> vec) : FullVector<T>(vec) {}
-		FullRow(int count) : FullVector<T>(count) {}
-		FullRow(int count, const T& value) : FullVector<T>(count, value) {}
+		FullRow(size_t count) : FullVector<T>(count) {}
+		FullRow(size_t count, const T& value) : FullVector<T>(count, value) {}
 		FullRow(typename std::vector<T>::const_iterator begin, typename std::vector<T>::const_iterator end) : FullVector<T>(begin, end) {}
 		FullRow(std::initializer_list<T> list) : FullVector<T>{ list } {}
 		FRowsptr<T> times(T a);
